% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{data_heatmap}
\alias{data_heatmap}
\title{Decomposition Heatmaps}
\usage{
data_heatmap(data, show_color_bar = TRUE, title = "", xlab = "", ylab = "")
}
\arguments{
\item{data}{List. The initial data blocks.}

\item{show_color_bar}{Boolean.}

\item{title}{Character.}

\item{xlab}{Character.}

\item{ylab}{Character}
}
\description{
Visualization of the RaJIVE decomposition, it shows heatmaps of the decomposition obtained by RaJIVE
}
