% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{decomposition_heatmaps_robustH}
\alias{decomposition_heatmaps_robustH}
\title{Decomposition Heatmaps}
\usage{
decomposition_heatmaps_robustH(blocks, jive_results_robust)
}
\arguments{
\item{blocks}{List. The initial data blocks.}

\item{jive_results_robust}{List. The RaJIVE decomposition.}
}
\value{
The heatmap of the decomposition
}
\description{
Visualization of the RaJIVE decomposition, it shows heatmaps of the decomposition obtained by RaJIVE
}
