% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{get_random_direction_bound_robustH}
\alias{get_random_direction_bound_robustH}
\title{Estimate the wedin bound for a data matrix.}
\usage{
get_random_direction_bound_robustH(n_obs, dims, num_samples = 1000)
}
\arguments{
\item{n_obs}{The number of observations.}

\item{dims}{The number of features in each data matrix}

\item{num_samples}{Integer. Number of vectors selected for resampling procedure.}
}
\value{
rand_dir_samples
}
\description{
Samples from the random direction bound. Returns on the scale of squared singular value.
}
