% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajive_helpfunctions.R
\name{get_svd}
\alias{get_svd}
\title{Computes the SVD of a matrix}
\usage{
get_svd(X, rank = NULL)
}
\arguments{
\item{rank}{number rank of SVD decomposition}

\item{decomposition}{X matrix.}
}
\value{
The SVD of X.
}
\description{
Computes the SVD of a matrix
}
