% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{get_svd_robustH}
\alias{get_svd_robustH}
\title{Computes the robust SVD of a matrix
Using robRsvd}
\usage{
get_svd_robustH(X, rank = NULL)
}
\arguments{
\item{X}{Matrix. X matrix.}

\item{rank}{Integer. Rank of SVD decomposition}
}
\value{
List. The SVD of X.
}
\description{
Computes the robust SVD of a matrix
Using robRsvd
}
