% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{showVarExplained_robust}
\alias{showVarExplained_robust}
\title{Proportions of variance explained}
\usage{
showVarExplained_robust(ajiveResults, blocks)
}
\arguments{
\item{ajiveResults}{List. The decomposition from Rajive}

\item{blocks}{List. The initial data blocks}
}
\value{
The proportion of variance explained by each component
}
\description{
Gets the variance explained by each component of the Rajive decomposition
}
