% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.functions.R
\name{sim_dist}
\alias{sim_dist}
\title{Simulation of single data block from distribution}
\usage{
sim_dist(num, n, p)
}
\arguments{
\item{num}{Integer. Type of distribution. 1 for normal, 2 for uniform, 3 for exponential}

\item{n}{Integer. Number of data points.}

\item{p}{Integers. Number of variables in  block.}
}
\description{
Simulation of single data block from distribution
}
