% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{decomposition_heatmaps_robustH}
\alias{decomposition_heatmaps_robustH}
\title{Decomposition Heatmaps}
\usage{
decomposition_heatmaps_robustH(blocks, jive_results_robust)
}
\arguments{
\item{blocks}{List. The initial data blocks.}

\item{jive_results_robust}{List. The RaJIVE decomposition.}
}
\description{
Visualization of the RaJIVE decomposition, it shows heatmaps of the decomposition obtained by RaJIVE
}
\examples{
blocks <- sample_toy_data(n=200, dx=100, dy=500)
initial_signal_ranks <- c(2, 2)
Rajive_decomp <- Rajive(blocks, initial_signal_ranks)
decomposition_heatmaps_robustH(blocks, Rajive_decomp)
}
