% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{get_block_loadings}
\alias{get_block_loadings}
\title{Block Loadings}
\usage{
get_block_loadings(ajive_output, k, type)
}
\arguments{
\item{ajive_output}{List. The decomposition from Rajive}

\item{k}{Integer. The index of the data block}

\item{type}{Character. Joint or individual}
}
\value{
The block loadings
}
\description{
Gets the block loadings from the Rajive decomposition
}
