% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{get_block_scores}
\alias{get_block_scores}
\title{Block Scores}
\usage{
get_block_scores(ajive_output, k, type)
}
\arguments{
\item{ajive_output}{List. The decomposition from Rajive}

\item{k}{Integer. The index of the data block}

\item{type}{Character. Joint or individual}

\item{normalized}{Boolean.}
}
\description{
Gets the block scores from the Rajive decomposition
}
