% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{get_individual_rank}
\alias{get_individual_rank}
\title{Individual Rank}
\usage{
get_individual_rank(ajive_output, k)
}
\arguments{
\item{ajive_output}{List. The decomposition from Rajive}

\item{k}{Integer. The index of the data block.}
}
\description{
Gets the individual ranks from the Rajive decomposition
}
