% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{get_joint_rank}
\alias{get_joint_rank}
\title{Joint Rank}
\usage{
get_joint_rank(ajive_output)
}
\arguments{
\item{ajive_output}{List. The decomposition from Rajive}
}
\description{
Gets the joint rank from the Rajive decomposition
}
