% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustSVD.R
\name{RobRSVD.all}
\alias{RobRSVD.all}
\title{Computes the robust SVD of a matrix}
\usage{
RobRSVD.all(data, nrank = min(dim(data)), svdinit = svd(data))
}
\arguments{
\item{data}{Matrix. X matrix.}

\item{nrank}{Integer. Rank of SVD decomposition}

\item{svdinit}{List. The standard SVD.}
}
\value{
List. The SVD of X.
}
\description{
Computes the robust SVD of a matrix
}
