% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rajive_helpfunctions.R
\name{truncate_svd}
\alias{truncate_svd}
\title{Truncates a robust SVD.}
\usage{
truncate_svd(decomposition, rank)
}
\arguments{
\item{decomposition}{List. List with entries 'u', 'd', and 'v'from the svd function.}

\item{rank}{List. List with entries 'u', 'd', and 'v'from the svd function.}
}
\value{
The trucated robust SVD of X.
}
\description{
Removes columns from the U, D, V matrix computed form an SVD.
}
